+++ oneadapter/src/androidTest/java/com/idanatz/oneadapter/tests/api/GetItemViewTypeFromClassMissingModuleException.kt
@@ -1,7 +1,7 @@
 package com.idanatz.oneadapter.tests.api
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import com.idanatz.oneadapter.external.UnsupportedClassException
+import com.idanatz.oneadapter.external.MissingModuleDefinitionException
 import com.idanatz.oneadapter.helpers.BaseTest
 import com.idanatz.oneadapter.models.TestModel
 import org.junit.Test
@@ -10,7 +10,7 @@ import org.junit.runner.RunWith
 @RunWith(AndroidJUnit4::class)
 class GetItemViewTypeFromClassMissingModuleException : BaseTest() {
 
-    @Test(expected = UnsupportedClassException::class)
+    @Test(expected = MissingModuleDefinitionException::class)
     fun test() {
         configure {
             act {