+++ oneadapter/src/main/java/com/idanatz/oneadapter/internal/holders/OneViewHolder.kt
@@ -7,38 +7,41 @@ import android.view.ViewGroup
 import androidx.annotation.LayoutRes
 import com.idanatz.oneadapter.internal.event_hooks.EventHooksMap
 import com.idanatz.oneadapter.external.event_hooks.SwipeEventHook
+import com.idanatz.oneadapter.external.interfaces.Diffable
 import com.idanatz.oneadapter.internal.selection.OneItemDetail
 import com.idanatz.oneadapter.internal.states.StatesMap
 import com.idanatz.oneadapter.internal.utils.extensions.inflateLayout
-import com.idanatz.oneadapter.internal.utils.extensions.let2
 
 @Suppress("NAME_SHADOWING")
-internal abstract class OneViewHolder<M> (
+internal abstract class OneViewHolder<M : Diffable> (
         parent: ViewGroup,
         @LayoutRes layoutResourceId: Int,
         val firstBindAnimation: Animator? = null,
-        private val statesHooksMap: StatesMap<M>? = null, // Not all ViewHolders will have states configured
-        private val eventsHooksMap: EventHooksMap<M>? = null // Not all ViewHolders will have events configured
+        val statesHooksMap: StatesMap<M>? = null, // Not all ViewHolders will have states configured
+        val eventsHooksMap: EventHooksMap<M>? = null // Not all ViewHolders will have events configured
 ) : RecyclerView.ViewHolder(parent.context.inflateLayout(layoutResourceId, parent, false)) {
 
     lateinit var viewBinder: ViewBinder
     lateinit var metadata: Metadata
-    var model: M? = null
+    lateinit var model: M
 
     // internal flags
     var isSwiping = false
 
     abstract fun onCreated()
-    abstract fun onBind(model: M?)
-    abstract fun onUnbind(model: M?)
+    abstract fun onBind(model: M)
+    abstract fun onUnbind(model: M)
+    abstract fun onClicked(model: M)
+    abstract fun onSwipe(canvas: Canvas, xAxisOffset: Float)
+    abstract fun onSwipeComplete(model: M, swipeDirection: SwipeEventHook.SwipeDirection)
+    abstract fun onSelected(model: M, selected: Boolean)
 
     fun onCreateViewHolder() {
         this.viewBinder = ViewBinder(itemView)
-        this.metadata = Metadata()
         onCreated()
     }
 
-    fun onBindViewHolder(model: M?, metadata: Metadata) {
+    fun onBindViewHolder(model: M, metadata: Metadata) {
         this.model = model
         this.metadata = metadata
 
@@ -47,9 +50,9 @@ internal abstract class OneViewHolder<M> (
         onBind(model)
     }
 
-    fun onBindSelection(model: M?, selected: Boolean) {
-        let2(statesHooksMap?.getSelectionState(), model) { selectionState, model ->
-            selectionState.onSelected(model, selected)
+    fun onBindSelection(model: M, selected: Boolean) {
+        if (statesHooksMap?.isSelectionStateConfigured() == true) {
+            onSelected(model, selected)
         }
     }
 
@@ -57,23 +60,19 @@ internal abstract class OneViewHolder<M> (
         if (isSwiping) // while swiping disable the option to select
             return null
 
-        return let2(statesHooksMap?.getSelectionState(), model) { selectionState, model ->
-            return if (selectionState.isSelectionEnabled(model)) {
-                OneItemDetail(adapterPosition, itemId)
-            } else
-                null
+        return statesHooksMap?.getSelectionState()?.let {
+            if (it.isSelectionEnabled(model)) OneItemDetail(adapterPosition, itemId)
+            else null
         }
     }
 
-    fun getSwipeEventHook() = eventsHooksMap?.getSwipeEventHook()
-
-    fun onSwipe(canvas: Canvas, xAxisOffset: Float) {
-        eventsHooksMap?.getSwipeEventHook()?.onSwipe(canvas, xAxisOffset, viewBinder)
+    fun onSwipeViewHolder(canvas: Canvas, xAxisOffset: Float) {
+        onSwipe(canvas, xAxisOffset)
     }
 
-    fun onSwipeComplete(swipeDirection: SwipeEventHook.SwipeDirection) {
-        let2(eventsHooksMap?.getSwipeEventHook(), model) { swipeEventHook, model ->
-            swipeEventHook.onSwipeComplete(model, swipeDirection, viewBinder)
+    fun onSwipeViewHolderComplete(swipeDirection: SwipeEventHook.SwipeDirection) {
+        if (eventsHooksMap?.isSwipeEventHookConfigured() == true) {
+            onSwipeComplete(model, swipeDirection)
         }
     }
 
@@ -87,8 +86,8 @@ internal abstract class OneViewHolder<M> (
     }
 
     private fun handleEventHooks() {
-        let2(eventsHooksMap?.getClickEventHook(), model) { clickEventHook, model ->
-            itemView.setOnClickListener { clickEventHook.onClick(model, viewBinder) }
+        if (eventsHooksMap?.isClickEventHookConfigured() == true) {
+            itemView.setOnClickListener { onClicked(model) }
         }
     }
 }