+++ oneadapter/src/main/java/com/idanatz/oneadapter/internal/states/StatesMap.kt
@@ -7,4 +7,5 @@ internal class StatesMap<M> : HashMap<String, State<M>>() {
 
     fun getSelectionState() = get(SelectionState.TAG) as SelectionState<M>?
     fun putSelectionState(state: SelectionState<M>) = put(SelectionState.TAG, state)
+    fun isSelectionStateConfigured() = getSelectionState() != null
 }