+++ oneadapter/src/main/java/com/idanatz/oneadapter/internal/threading/OneSingleThreadPoolExecutor.kt
@@ -6,6 +6,18 @@ import java.util.concurrent.*
 
 internal class OneSingleThreadPoolExecutor : ThreadPoolExecutor(1, 1, 0L, TimeUnit.MILLISECONDS, LinkedBlockingQueue()) {
 
+    fun submit(taskId: Int, task: Runnable): IdentifiableFuture<*> {
+        val future = super.submit(task)
+        return object : IdentifiableFuture<Any> {
+            override val id: Int = taskId
+            override fun isDone(): Boolean = future.isDone
+            override fun get() = future.get()
+            override fun get(timeout: Long, unit: TimeUnit?) = future.get(timeout, unit)
+            override fun cancel(mayInterruptIfRunning: Boolean) = future.cancel(mayInterruptIfRunning)
+            override fun isCancelled() = future.isCancelled
+        }
+    }
+
     override fun afterExecute(r: Runnable?, t: Throwable?) {
         super.afterExecute(r, t)
 
@@ -24,4 +36,8 @@ internal class OneSingleThreadPoolExecutor : ThreadPoolExecutor(1, 1, 0L, TimeUn
             }
         }
     }
+}
+
+internal interface IdentifiableFuture<V : Any> : Future<V> {
+    val id: Int
 }