+++ oneadapter/src/main/java/com/idanatz/oneadapter/internal/utils/Logger.kt
@@ -4,9 +4,9 @@ import android.util.Log
 import com.idanatz.oneadapter.BuildConfig
 import com.idanatz.oneadapter.internal.InternalAdapter
 
-internal object Logger {
+internal class Logger(val internalAdapter: InternalAdapter) {
 
-    fun logd(internalAdapter: InternalAdapter? = null, body: () -> String) {
+    fun logd(body: () -> String) {
         if (BuildConfig.DEBUG) {
             Log.d("OneAdapter", "adapter: $internalAdapter, ${body()}")
         }